#62 Moved tests
